pyproject.toml
[project]
name = "LSMS_Library"
dynamic = ["version"]
description = "Abstraction layer for Living Standards Measurement Survey data"
authors = [ {name = "Ethan Ligon", email = "ligon@berkeley.edu"}
    ]
license = {text = "CC-BY-4.0"}
readme = "README.org"
requires-python = ">=3.11, <4.0"
dependencies = ["pandas >=2.2",
             "CFEDemands>=0.8.2",
             "PyYAML >=6.0,<7.0",
             "python-gnupg >=0.5,<1.0",
             "scikit-learn (>=1.7.0,<2.0.0)",
             "dvc[s3] (>=3.67.0,<4.0.0)",
             "pyarrow (>=21.0.0)",
             "pyreadstat (>=1.3.0,<2.0.0)",
             "openai (>=1.99.5,<2.0.0)",
            "ligonlibrary (>=0.2.0)",
             "python-Levenshtein (>=0.27.3,<0.28.0)",
             "typer (>=0.12.0,<1.0.0)",
             "fooddatacentral (>=1.0.10,<2.0.0)",
             "pint (>=0.25.2,<0.26.0)",
             "platformdirs (>=4.0.0,<5.0.0)"]

[project.scripts]
lsms-library = "lsms_library.cli:main"

[project.urls]
repository = "https://github.com/ligon/LSMS_Library"

[tool.poetry]
include = [
    "lsms_library/countries/.dvc/config",
    "lsms_library/countries/.dvc/s3_config",
    "lsms_library/countries/.dvc/.gitignore",
    "lsms_library/countries/.dvc/plots/*.json",
    "lsms_library/countries/.dvc/s3_reader_creds.gpg",
]
version = "0.0.0"

[tool.poetry.requires-plugins]
poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
pytest-xdist = "^3.6.0"
pytest-instafail = "^0.5.0"
ipython = "^9.4.0"
matplotlib = "^3.10.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs-material = ">=9.0"
mkdocstrings = {extras = ["python"], version = ">=0.28"}

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[build-system]
requires = ["poetry-core>=2.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
